setting up the function arborescence for the analysis.
main function is analyze() and is called with a list of files to analyze as argument. the function gets the information of every file by stripping their names (function getinfo() ) and creates the appropriate folder to store the results (function makefolder() ), then calls for one of three macro algorithms depending on the user selection. the functions still have to be written but are placed in the loop